move spinner to separate file, fix comments
