+++ static/js/apps/biomed_nl/app.tsx
@@ -22,13 +22,7 @@
 import { css, ThemeProvider } from "@emotion/react";
 import axios from "axios";
 import queryString from "query-string";
-import React, {
-  ReactElement,
-  useEffect,
-  useMemo,
-  useRef,
-  useState,
-} from "react";
+import React, { ReactElement, useEffect, useState } from "react";
 import Collapsible from "react-collapsible";
 import ReactMarkdown from "react-markdown";
 import { Input, InputGroup } from "reactstrap";
@@ -41,11 +35,36 @@ import { Link, LinkBox } from "../../components/elements/link_box";
 import theme from "../../theme/theme";
 import { stringifyFn } from "../../utils/axios";
 import { updateHash } from "../../utils/url_utils";
+import { SpinnerWithText } from "./spinner";
 
 // Constants for URL hash parameters
 const URL_HASH_PARAMS = {
   q: "q",
 };
+
+const sampleQuestions = [
+  "What genes are associated with Alzheimer's disease?",
+  "What is the mechanism of action of atorvastatin?",
+  "Is there a complete genome for Monkeypox virus?",
+  "What genes does rs429358 regulate?",
+  "What is the hg38 genomic location of FGFR1?",
+  "What type of gene is FGFR1?",
+  "What is the ICD10 code of meningococcal meningitis?",
+  "What is the UMLS CUI for rheumatoid arthritis?",
+  "What can you tell me about atorvastatin?",
+  "What genetic variants are associated with premature birth?",
+  "What is the host of Betapapillomavirus 1?",
+  "What is the HGNC ID of FGFR1?",
+  "What is the concept unique id for rheumatoid arthritis?",
+  "What is the chembl ID of acetaminophen?",
+];
+const overviewText = `This experiment allows you to explore the Biomedical Data
+  Commons knowledge graph with ease using natural language. Ask
+  research questions in natural language, like "What genes are
+  associated with Alzheimer's disease?". The AI model will
+  interpret your query, search the knowledge graph, and return
+  concise answers with links to relevant data. Your feedback is
+  invaluable; a feedback form will appear after each query.`;
 
 // Interface for the response received from Biomed NL API.
 interface BiomedNlApiResponse {
@@ -60,7 +79,7 @@ interface DisplayedAnswer {
   debugInfo: string;
 }
 
-// Eneable dropdown for Footer and Debug sections
+// Headings for Footer and Debug dropdown sections
 function getSectionTrigger(title: string, opened: boolean): JSX.Element {
   return (
     <div
@@ -102,79 +121,6 @@ export function App(): ReactElement {
   const [answer, setAnswer] = useState<DisplayedAnswer>(null);
   const [showLoading, setShowLoading] = useState<boolean>(false);
 
-  // A loading spinner that shows new biomed pun every minute
-  function SpinnerWithText(): ReactElement {
-    const [showMessage, setShowMessage] = useState(false);
-    const [message, setMessage] = useState("");
-    const [messageIndex, setMessageIndex] = useState(0);
-
-    const customMessages = useMemo(
-      () => [
-        "The data is taking its time to ensure perfect homeostasis.",
-        "Like a healthy cell, the data is going through a natural maturation process.",
-        "The data is undergoing a controlled differentiation, becoming more specialized.",
-        "The data is percolating through the system, ensuring complete absorption.",
-        "The data is taking its time to allow for complete osmosis of information.",
-        "The data is making sure all its enzymes are perfectly aligned for the reaction.",
-      ],
-      []
-    );
-    // Show puns in random order every time.
-    const shuffledIndexes = useRef(
-      [...Array(customMessages.length).keys()].sort(() => Math.random() - 0.5)
-    ).current;
-
-    useEffect(() => {
-      // Show first biomed pun after 3 seconds
-      const timeoutId = setTimeout(() => {
-        setShowMessage(true);
-        setMessage(customMessages[shuffledIndexes[messageIndex]]);
-      }, 3000);
-
-      // Show next pun after one minute
-      const messageInterval = setInterval(() => {
-        setMessageIndex((prevIndex) => (prevIndex + 1) % customMessages.length);
-        setMessage(customMessages[shuffledIndexes[messageIndex]]);
-      }, 60000);
-
-      return () => {
-        clearTimeout(timeoutId);
-        clearInterval(messageInterval);
-      };
-    }, [customMessages, messageIndex, shuffledIndexes]);
-
-    return (
-      <>
-        <div className="spinner-container">
-          <div className="spinner"></div>
-          <div>This may take up to 5 minutes...</div>
-        </div>
-        {showMessage && (
-          <div
-            css={css`
-              margin: ${theme.spacing.xl}px;
-            `}
-          >
-            <p
-              css={css`
-                ${theme.typography.text.md};
-              `}
-            >
-              {message}*
-            </p>
-            <p
-              css={css`
-                ${theme.typography.text.sm};
-              `}
-            >
-              * Puns generated by Gemini
-            </p>
-          </div>
-        )}
-      </>
-    );
-  }
-
   /**
    * useEffect hook to handle initial loading of information from the URL hash.
    */
@@ -206,7 +152,6 @@ export function App(): ReactElement {
   function handleKeydownEvent(
     event: React.KeyboardEvent<HTMLDivElement>
   ): void {
-    // Navigate through the suggested results.
     switch (event.key) {
       case "Enter":
         event.preventDefault();
@@ -240,30 +185,6 @@ export function App(): ReactElement {
         setShowLoading(false);
       });
   }
-
-  const sampleQuestions = [
-    "What genes are associated with Alzheimer's disease?",
-    "What is the mechanism of action of atorvastatin?",
-    "Is there a complete genome for Monkeypox virus?",
-    "What genes does rs429358 regulate?",
-    "What is the hg38 genomic location of FGFR1?",
-    "What type of gene is FGFR1?",
-    "What is the ICD10 code of meningococcal meningitis?",
-    "What is the UMLS CUI for rheumatoid arthritis?",
-    "What can you tell me about atorvastatin?",
-    "What genetic variants are associated with premature birth?",
-    "What is the host of Betapapillomavirus 1?",
-    "What is the HGNC ID of FGFR1?",
-    "What is the concept unique id for rheumatoid arthritis?",
-    "What is the chembl ID of acetaminophen?",
-  ];
-  const overviewText = `This experiment allows you to explore the Biomedical Data
-    Commons knowledge graph with ease using natural language. Ask
-    research questions in natural language, like "What genes are
-    associated with Alzheimer's disease?". The AI model will
-    interpret your query, search the knowledge graph, and return
-    concise answers with links to relevant data. Your feedback is
-    invaluable; a feedback form will appear after each query.`;
 
   return (
     <ThemeProvider theme={theme}>
